+++ R/fit1.R
@@ -87,7 +87,7 @@
 #' out_fit1 <- fit1(probs[[7]][,,max_pos], pheno, addcovar=covar)
 #'
 #' # fit QTL model just at that position, with polygenic effect
-#' # out_fit1 <- fit1(probs[[7]][,,max_pos], pheno, kinship7, addcovar=covar)
+#' out_fit1_pg <- fit1(probs[[7]][,,max_pos], pheno, kinship7, addcovar=covar)
 #'
 #' @importFrom stats setNames
 #' @export